moving link functions to utils
